+++ toshi_hazard_store/model/pyarrow/dataset_schema.py
@@ -2,12 +2,15 @@
 Define the standard schema used with pyarrow datasets.
 """
 
-import pyarrow
+import pyarrow as pa
+from lancedb.pydantic import pydantic_to_schema
+
+from toshi_hazard_store.model.hazard_models_pydantic import HazardAggregateCurve
 
 USE_64BIT_VALUES_DEFAULT = False
 
 
-def get_hazard_realisation_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEFAULT) -> pyarrow.schema:
+def get_hazard_realisation_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEFAULT) -> pa.schema:
     """A schema for Hazard Realization curves dataset extracted from openquake.
 
     Attributes:
@@ -25,13 +28,13 @@ def get_hazard_realisation_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEF
         values: a list of the 44 IMTL values
     """
     # create a schema...
-    vtype = pyarrow.float64() if use_64_bit_values else pyarrow.float32()
-    values_type = pyarrow.list_(vtype)
-    vs30_type = pyarrow.int32()
-    dict_type = pyarrow.dictionary(pyarrow.int8(), pyarrow.string(), False)
-    str_type = pyarrow.string()
+    vtype = pa.float64() if use_64_bit_values else pa.float32()
+    values_type = pa.list_(vtype)
+    vs30_type = pa.int32()
+    dict_type = pa.dictionary(pa.int8(), pa.string(), False)
+    str_type = pa.string()
 
-    return pyarrow.schema(
+    return pa.schema(
         [
             ("compatible_calc_id", str_type),  # for hazard-calc equivalence, for PSHA engines interoperability
             ("producer_digest", dict_type),  # digest for the producer look up
@@ -49,41 +52,20 @@ def get_hazard_realisation_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEF
     )
 
 
-def get_hazard_aggregate_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEFAULT) -> pyarrow.schema:
+def get_hazard_aggregate_schema(use_64_bit_values: bool = USE_64BIT_VALUES_DEFAULT) -> pa.schema:
     """A schema for aggregate hazard curve datasets.
 
-    Generally these are aggregated realisation curves.
-
-    Attributes:
-        compatible_calc_id: for hazard-calc equivalence.
-        model_id: the model that these curves represent.
-        nloc_001:  the location string to three places e.g. "-38.330~17.550".
-        nloc_0:  the location string to zero places e.g.  "-38.0~17.0" (used for partitioning).
-        imt:  the imt label e.g. 'PGA', 'SA(5.0)'
-        vs30:  the VS30 integer
-        aggr: the aggregation type
-        values: a list of the 44 IMTL values
+    built dynamically from the pydantic model, using lancedb helper method.
     """
-    # create a schema...
 
-    vtype = pyarrow.float64() if use_64_bit_values else pyarrow.float32()
-    values_type = pyarrow.list_(vtype)
+    # Convert the Pydantic model to a PyArrow schema
+    arrow_schema = pydantic_to_schema(HazardAggregateCurve)
+    if not use_64_bit_values:
+        arrow_schema = arrow_schema.set(
+            arrow_schema.get_field_index('vs30'), pa.lib.field('vs30', pa.int32(), nullable=False)
+        )
+        arrow_schema = arrow_schema.set(
+            arrow_schema.get_field_index('values'), pa.lib.field('values', pa.list_(pa.float32()), nullable=False)
+        )
 
-    vs30_type = pyarrow.int32()
-
-    # # dict_type = pyarrow.string()
-    # dict_type_int8 = pyarrow.dictionary(pyarrow.int8(), pyarrow.string())
-    # # dict_type_int16 = pyarrow.dictionary(pyarrow.int16(), pyarrow.string())
-
-    return pyarrow.schema(
-        [
-            ("compatible_calc_id", pyarrow.string()),  # for hazard-calc equivalence, for PSHA engines interoperability
-            ("hazard_model_id", pyarrow.string()),  # the model that these curves represent.
-            ("nloc_001", pyarrow.string()),  # the location string to three places e.g. "-38.330~17.550"
-            ("nloc_0", pyarrow.string()),  # the location string to zero places e.g.  "-38.0~17.0" (used for partioning)
-            ('imt', pyarrow.string()),  # the imt label e.g. 'PGA', 'SA(5.0)'
-            ('vs30', vs30_type),  # the VS30 integer
-            ('aggr', pyarrow.string()),  # the the aggregation type
-            ("values", values_type),  # a list of the 44 IMTL values
-        ]
-    )
+    return arrow_schema